refactor toolsSetup
